Move project_conversation tools to the project_manager and sunset the old server.
