+++ front/lib/actions/mcp_internal_actions/constants.test.ts
@@ -67,7 +67,6 @@ describe("INTERNAL_MCP_SERVERS", () => {
       { name: "agent_sidekick_context", id: 1022 },
       { name: "agent_sidekick_agent_state", id: 1023 },
       { name: "sandbox", id: 1024 },
-      { name: "project_conversation", id: 1025 },
       { name: "user_mentions", id: 1026 },
       { name: "ask_user_question", id: 1028 },
     ];
