+++ front/lib/resources/skill/global/projects.ts
@@ -68,7 +68,7 @@ When you need to find information, uses this order (skip steps if the relevant t
 
     return instructions;
   },
-  mcpServers: [{ name: "project_manager" }, { name: "project_conversation" }],
+  mcpServers: [{ name: "project_manager" }],
   version: 1,
   icon: "ActionFolderIcon",
   isRestricted: async (auth: Authenticator) => {
